Add files via upload
DicePlay - board game in which each player gets to choose a pawn to play with. The players must roll the dice one by one, then pick up a card that impacts their score in either a positive or a negative way. The first player to get to the finish line is declared the winner of the game. The game was made in Unity, using C# for the back-end. The file "Fisiere C#" contains all the scripts that compose the game functionalities.